requirements.txt
pytest
requests
selenium
webdriver-manager
argon2-cffi
pytest-html
pytest-metadata
allure-pytest
python-dotenv
boto3
